Move escape methods into own utility, add unit tests
